Refactor: separate api call mutual funds
