+++ frontend/src/components/CalculateButton.js
@@ -1,7 +1,11 @@
 import React from "react";
 
 const CalculateButton = ({ handleSubmit }) => {
-  return <button type="submit" onClick={handleSubmit}>Calculate Future Value</button>;
+  return (
+    <button type="submit" onClick={handleSubmit}>
+      Calculate Future Value
+    </button>
+  );
 };
 
 export default CalculateButton;
